requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
appdirs==1.4.4
asynctest==0.13.0
attributee==0.1.9
attrs==24.2.0
backcall==0.2.0
bypy==1.8.9
cachetools==5.5.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.4.2
cityscapesScripts==2.2.4
click==8.1.8
clip==1.0
cloudpickle==2.2.1
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.2.7
crcmod==1.7
cryptography==45.0.6
cycler==0.11.0
Cython==3.0.12
decorator==5.1.1
dill==0.3.7
dotty-dict==1.3.1
einops==0.6.1
exceptiongroup==1.3.0
fairscale==0.4.6
filelock==3.12.2
flake8==5.0.4
flatbuffers==25.2.10
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
google-auth==2.40.3
google-auth-oauthlib==0.4.6
grpcio==1.62.3
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.10
imagecorruptions==1.1.2
imageio==2.31.2
importlib-metadata==6.7.0
iniconfig==2.0.0
interrogate==1.5.0
iopath==0.1.10
ipdb==0.13.13
ipython==7.34.0
isort==4.3.21
jedi==0.19.2
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
kwarray==0.6.20
lap==0.5.12
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.7.0
mmdet==2.28.2
mmengine==0.10.7
-e git+https://github.com/open-mmlab/mmtracking@e79491ec8f0b8c86fda947fbaaa824c66ab2a991#egg=mmtrack
model-index==0.1.11
motmetrics==1.4.0
mpmath==1.3.0
multiprocess==0.70.15
MultiScaleDeformableAttention==1.0
networkx==2.6.3
ninja==1.11.1.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.7.0
onnxruntime==1.14.1
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.0.0
pluggy==1.2.0
portalocker==2.7.0
prompt_toolkit==3.0.48
protobuf==3.20.1
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.23.0
pyflakes==2.5.0
Pygments==2.17.2
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyWavelets==1.3.0
PyYAML==6.0.1
regex==2024.4.16
requests==2.28.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.17.0
submitit==1.4.6
sympy==1.10.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.12
toml==0.10.2
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.65.2
traitlets==5.9.0
typing==3.7.4.3
typing_extensions==4.7.1
ubelt==1.3.8
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==2.2.3
xdoctest==1.1.6
xmltodict==0.14.2
yacs==0.1.8
yapf==0.40.1
zipp==3.15.0
